requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-openai==0.1.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10
newspaper3k==0.2.8